bot/requirements.txt
aiogram==3.13.1
aiogram-dialog==2.3.0
httpx==0.27.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0



